Fix shader debugging and textures
Shader debugging is a little easier as it will try to tell you which file
has the error now. (When compiling, at least. It's rather hard to say
where the issue is at the link step. :( It might work well if I list all
of the files in the linker?)