Add automatic clang-format (fork of v8.0.0)
There is now a custom version of clang-format that implements the
DanglingParenthesis option. Furthermore, I have added hooks and
vscode settings changes to automatically format any C++ code. Some
changes were made to the style guide to deal with minor problems with
clang-format as-is, however, future revisions will be made to the fork
to make these automated as well as simplifying the build of clang-format
to be fully standalone, rather than requiring to build as significant
portion of LLVM. I may (at some indefinite future point) create a
custom lexer and functionally extensible yet opinionated C++ formatter
(in line with Prettier or `rustfmt`).

Many surface changes were applied throughout the code, as per the
automation.